1. Priority scheme is slightly changed: do everything certain first, then move to delayed jobs. 2. Driver now aware of time elapsed after last "application stop" event. Android system requires some amount of time to clear up previously died application instance.

+++ src/edu/berkeley/wtchoi/cc/util/E.java
@@ -12,7 +12,7 @@ package edu.berkeley.wtchoi.cc.util;
 public abstract class E {
     private static int miniSleepInterval = 200;
 
-    public static boolean sleep(int sleepInterval){
+    public static boolean sleep(long sleepInterval){
         try{
             Thread.sleep(sleepInterval);
             return true;
